+++ src/production.rs
@@ -2,4 +2,5 @@ mod co_class;
 #[doc(hidden)]
 pub mod registration;
 
+#[doc(inline)]
 pub use co_class::CoClass;
